Refactor [#9] Category 관련 패키지 이동
application으로 Service 이동
